Split of store into modules for easier code maintenance
